rename: add about folder
