+++ components/about/keywords/KeywordsRotatingText.tsx
