--- components/career/CareerItem.tsx
@@ -1,37 +1,0 @@
-import styled from "styled-components";
-
-const CareerItem = ({
-  period,
-  content,
-}: Pick<Career, "content" | "period">) => {
-  return (
-    <Item>
-      <Period>{period}</Period>
-      <Content>{content}</Content>
-    </Item>
-  );
-};
-
-export default CareerItem;
-
-const Item = styled.li`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  border-bottom: 1px solid lightgray;
-  padding: 1rem 0.6rem;
-  gap: 15px;
-  h3 {
-    font-size: 1.3rem;
-  }
-`;
-
-const Period = styled.p`
-  font-size: 1rem;
-  color: #828282;
-`;
-
-const Content = styled.p`
-  text-align: center;
-  font-size: 1.2rem;
-`;
